+++ hypervisor/dm/vpci/vdev.c
@@ -28,7 +28,7 @@
  */
 
 #include <vm.h>
-#include "pci_priv.h"
+#include "vpci_priv.h"
 
 uint32_t pci_vdev_read_cfg(const struct pci_vdev *vdev, uint32_t offset, uint32_t bytes)
 {
