+++ hypervisor/dm/vpci/vpci_priv.h
